[refactor] #81 : 파일 이름 수정
