+++ src/features/find/ui/LocationStep.tsx
@@ -6,7 +6,7 @@ import { FormattedData, StartPointInfo } from "../model";
 import { highlightMatchingText } from "@/shared/utils";
 import { useSearchParams } from "react-router-dom";
 import { useCreateStartPoint, useSearch } from "../hooks";
-import { PlainHeader } from "@/widgets/Header";
+import { PlainHeader } from "@/widgets/headers";
 
 interface StartPoint {
   id: string;
